# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenAPI",
        "description": "Laravel Front OpenApi",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/cms/tags/{slug}": {
            "get": {
                "tags": [
                    "cms"
                ],
                "summary": "Tag detail",
                "description": "Tag detail",
                "operationId": "App\\Api\\Controllers\\TagController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/slug"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
...
            "slug": {
                "name": "slug",
                "in": "path",
                "description": "Identifiant URL",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    },
...
}